refactor: extract 4 kernel test modules into integration tests (#825)
Extract egress_tests, receipt_chain_tests, kernel_token_verification_tests,
and sink_scope_tests from kernel.rs (5,072 lines) into separate integration
test files under crates/portcullis/tests/. Reduces kernel.rs to 4,297 lines
(775 lines removed, all 33 tests preserved).

- kernel_egress.rs: egress policy host allow/deny tests (requires "spec")
- kernel_receipt.rs: receipt chain hash-linking tests
- kernel_token.rs: Ed25519 declassification token verification (requires "crypto")
- kernel_sink_scope.rs: certificate sink scope enforcement tests
- Add VerifiedPermissions::new() constructor for non_exhaustive struct access
- Lower .line-ratchet.toml ceiling from 5100 to 4300
