+++ .line-ratchet.toml
@@ -20,6 +20,6 @@ path = "crates/nucleus-claude-hook/src/main.rs"
 
 [[files]]
 path = "crates/portcullis/src/kernel.rs"
-ceiling = 5100
+ceiling = 4300
 target = 2500
 step = 1
